refactor(approve-intent): extract IntentDetailWithApproval with runtime type guard
