+++ frontend/src/features/approve-intent/index.ts
@@ -2,3 +2,4 @@ export { useApproveIntent } from "./api/useApproveIntent";
 export type { IntentApprovalStatus, IntentApprovalAction } from "./model/types";
 export { ApproveIntentDialog } from "./ui/ApproveIntentDialog";
 export { IntentStatusControl } from "./ui/IntentStatusControl";
+export { IntentDetailWithApproval } from "./ui/IntentDetailWithApproval";
