+++ frontend/src/pages/domain-pack/ui/IntentDraftReadPage.tsx
@@ -1,90 +1,9 @@
-import { useEffect, useState } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 import { IntentDetailPanel, IntentTreePanel } from "../../../features/intent-draft-read/ui";
-import {
-  useApproveIntent,
-  type IntentApprovalStatus,
-  type IntentApprovalAction,
-  ApproveIntentDialog,
-  IntentStatusControl,
-} from "../../../features/approve-intent";
+import { IntentDetailWithApproval } from "../../../features/approve-intent";
 import { parseRouteId } from "../../../shared/lib/parseRouteId";
 import { DashboardLayout } from "../../../shared/ui/layout/DashboardLayout";
 import styles from "./intent-draft-read-page.module.css";
-
-function IntentDetailWithApproval({
-  wsId,
-  pId,
-  vId,
-  iId,
-}: {
-  wsId: number;
-  pId: number;
-  vId: number;
-  iId: number;
-}) {
-  const [dialogAction, setDialogAction] = useState<IntentApprovalAction | null>(null);
-  const [statusOverride, setStatusOverride] = useState<string | null>(null);
-  const [detailRefreshKey, setDetailRefreshKey] = useState(0);
-
-  useEffect(() => {
-    setDialogAction(null);
-    setStatusOverride(null);
-    setDetailRefreshKey(0);
-  }, [iId]);
-
-  const mutation = useApproveIntent({
-    wsId,
-    packId: pId,
-    versionId: vId,
-    intentId: iId,
-    onStatusChanged: (status) => {
-      setStatusOverride(status);
-      setDialogAction(null);
-      setDetailRefreshKey((k) => k + 1);
-    },
-  });
-
-  const handleDialogOpenChange = (next: boolean) => {
-    if (!mutation.isPending) setDialogAction(next ? dialogAction : null);
-  };
-
-  const handleConfirm = () => {
-    if (dialogAction === null) return;
-    const status: IntentApprovalStatus =
-      dialogAction === "publish" ? "PUBLISHED" : "REJECTED";
-    mutation.mutate(status);
-  };
-
-  return (
-    <IntentDetailPanel
-      wsId={wsId}
-      packId={pId}
-      versionId={vId}
-      intentId={iId}
-      refreshKey={detailRefreshKey}
-    >
-      {(detail) => (
-        <>
-          <IntentStatusControl
-            intentStatus={(statusOverride ?? detail.status) as "DRAFT" | IntentApprovalStatus}
-            onPublish={() => setDialogAction("publish")}
-            onReject={() => setDialogAction("reject")}
-            isPending={mutation.isPending}
-          />
-          <ApproveIntentDialog
-            intentName={detail.name}
-            action={dialogAction ?? "publish"}
-            open={dialogAction !== null}
-            onOpenChange={handleDialogOpenChange}
-            onConfirm={handleConfirm}
-            isLoading={mutation.isPending}
-          />
-        </>
-      )}
-    </IntentDetailPanel>
-  );
-}
 
 export function IntentDraftReadPage() {
   const { workspaceId, packId, versionId, intentId } = useParams();
@@ -148,7 +67,7 @@ export function IntentDraftReadPage() {
           </div>
           <div className={styles.detailSlot}>
             {iId !== null ? (
-              <IntentDetailWithApproval wsId={wsId} pId={pId} vId={vId} iId={iId} />
+              <IntentDetailWithApproval key={iId} wsId={wsId} pId={pId} vId={vId} iId={iId} />
             ) : (
               <IntentDetailPanel wsId={wsId} packId={pId} versionId={vId} intentId={null} />
             )}
